Refactored filetracker client into a separate package.
This commit breaks up huge filetracker/__init__.py, that
contains the client code from the times when there was no server,
into logical parts inside the new filetracker.client package.

No logic should be touched in this commit except 'DataStore.get_file',
which was decoupled from LocalDataStore.get_file and its hardlinking
semantics.
